+++ package.json
@@ -11,9 +11,11 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "axios": "^0.27.2",
     "cookie-session": "^2.0.0",
     "express": "^4.17.2",
     "helmet": "^5.1.1",
+    "mongoose": "^6.5.2",
     "morgan": "^1.10.0",
     "passport": "^0.5.0",
     "passport-github2": "^0.1.12",
@@ -21,8 +23,10 @@
     "typescript": "^4.5.5"
   },
   "devDependencies": {
+    "@types/axios": "^0.14.0",
     "@types/cookie-session": "^2.0.44",
     "@types/express": "^4.17.13",
+    "@types/mongoose": "^5.11.97",
     "@types/morgan": "^1.9.3",
     "@types/node": "^17.0.10",
     "@types/passport": "^1.0.10",
