+++ src/app.ts
@@ -4,56 +4,13 @@ import morgan from 'morgan';
 import passport from 'passport';
 import path from 'path';
 import auth from './router/auth.router';
-import { COOKIE_SECRET, GITHUB_KEYS, GOOGLE_KEYS } from './constants';
-import passportGoogle from 'passport-google-oauth20';
-import passportGithub from 'passport-github2';
+import { COOKIE_SECRET } from './constants';
 import cookieSession from 'cookie-session';
 import { isLoggedIn, isNotLoggedIn } from './middleware/checkAuth';
+import { googleStrategy } from './middleware/GoogleStrategy';
+import { githubStrategy } from './middleware/GithubStrategy';
 
 const app = express();
-
-passport.use(
-  new passportGoogle.Strategy(
-    {
-      clientID: GOOGLE_KEYS.CLIENT_ID,
-      clientSecret: GOOGLE_KEYS.CLIENT_SECRET,
-      callbackURL: '/auth/google/callback',
-    },
-    verifyCallback,
-  ),
-);
-
-function verifyCallback(
-  accessToken: any,
-  refreshToken: any,
-  profile: any,
-  done: any,
-) {
-  console.log('Success in verify');
-  console.log(profile);
-  done(null, profile);
-}
-
-passport.use(
-  new passportGithub.Strategy(
-    {
-      clientID: GITHUB_KEYS.CLIENT_ID,
-      clientSecret: GITHUB_KEYS.CLIENT_SECRET,
-      callbackURL: '/auth/github/callback',
-    },
-    verifyCallback2,
-  ),
-);
-
-function verifyCallback2(
-  accessToken: any,
-  refreshToken: any,
-  profile: any,
-  done: any,
-) {
-  console.log(profile.username);
-  done(null, profile.username);
-}
 
 passport.serializeUser((user: any, done) => {
   console.log('Serialize User');
@@ -63,9 +20,9 @@ passport.serializeUser((user: any, done) => {
 
 passport.deserializeUser((user: any, done) => {
   console.log('Deserialize User');
-  console.log(user.displayName);
+  // Will Extract the Cookie
   // Assigning a value to req.user
-  done(null, { ad: 'sample' });
+  done(null, user);
 });
 
 app.use(
@@ -89,6 +46,9 @@ app.use(
     keys: [COOKIE_SECRET.COOKIE_SECRET1, COOKIE_SECRET.COOKIE_SECRET2],
   }),
 );
+passport.use(googleStrategy());
+passport.use(githubStrategy());
+
 app.use(passport.initialize());
 
 // Allows serialization and deserialization
